// frontend/src/app/generated/model/recipeIngredientDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RecipeIngredientDto { 
    ingredientId?: number;
    name?: string;
    type?: RecipeIngredientDto.TypeEnum;
    quantity?: number;
    unit?: RecipeIngredientDto.UnitEnum;
}
export namespace RecipeIngredientDto {
    export type TypeEnum = 'VEGETABLE' | 'MEAT' | 'MILK_PRODUCT' | 'SPICE' | 'SALT_GROCERY' | 'SWEET_GROCERY' | 'FRUIT' | 'CONDIMENT';
    export const TypeEnum = {
        Vegetable: 'VEGETABLE' as TypeEnum,
        Meat: 'MEAT' as TypeEnum,
        MilkProduct: 'MILK_PRODUCT' as TypeEnum,
        Spice: 'SPICE' as TypeEnum,
        SaltGrocery: 'SALT_GROCERY' as TypeEnum,
        SweetGrocery: 'SWEET_GROCERY' as TypeEnum,
        Fruit: 'FRUIT' as TypeEnum,
        Condiment: 'CONDIMENT' as TypeEnum
    };
    export type UnitEnum = 'KG' | 'G' | 'L' | 'ML' | 'PACKET' | 'NA' | 'PINCH' | 'TABLESPOON' | 'TEASPOON';
    export const UnitEnum = {
        Kg: 'KG' as UnitEnum,
        G: 'G' as UnitEnum,
        L: 'L' as UnitEnum,
        Ml: 'ML' as UnitEnum,
        Packet: 'PACKET' as UnitEnum,
        Na: 'NA' as UnitEnum,
        Pinch: 'PINCH' as UnitEnum,
        Tablespoon: 'TABLESPOON' as UnitEnum,
        Teaspoon: 'TEASPOON' as UnitEnum
    };
}